Move AddressComboBox to separate component
