+++ apps/public/src/app/(map)/locatie/kies/_components/SideBar/SideBar.test.tsx
@@ -1,18 +1,6 @@
 import { render, screen, waitFor } from '@testing-library/react'
-import userEvent from '@testing-library/user-event'
-import { http, HttpResponse } from 'msw'
-import { vi } from 'vitest'
-
-import { ENDPOINTS } from 'apps/public/src/mocks/endpoints'
-import { server } from 'apps/public/src/mocks/node'
 
 import { SideBar } from './SideBar'
-
-global.ResizeObserver = vi.fn().mockImplementation(() => ({
-  observe: vi.fn(),
-  unobserve: vi.fn(),
-  disconnect: vi.fn(),
-}))
 
 describe('SideBar', () => {
   it('should render correctly', () => {
@@ -25,101 +13,11 @@ describe('SideBar', () => {
     expect(paragraph).toBeInTheDocument()
   })
 
-  it('should render render address based on provided coordinates ', async () => {
+  it('should show an address based on provided coordinates ', async () => {
     render(<SideBar coordinates={{ lat: 52.37239126063553, lon: 4.900905743712159 }} />)
 
     await waitFor(() => {
       expect(screen.getByDisplayValue('Nieuwmarkt 15, 1011JR Amsterdam')).toBeInTheDocument()
-    })
-  })
-
-  it('should render the address input', () => {
-    render(<SideBar />)
-
-    const input = screen.getByRole('combobox', { name: 'Zoek op adres' })
-
-    expect(input).toBeInTheDocument()
-  })
-
-  it('should not show the list box initially', () => {
-    render(<SideBar />)
-
-    const listBox = screen.queryByRole('listbox')
-
-    expect(listBox).not.toBeInTheDocument()
-  })
-
-  it('should not show the list box on 2 character input', async () => {
-    const user = userEvent.setup()
-
-    render(<SideBar />)
-
-    const input = screen.getByRole('combobox', { name: 'Zoek op adres' })
-
-    await user.type(input, 'aa')
-
-    await waitFor(() => {
-      const listBox = screen.queryByRole('listbox')
-      expect(listBox).not.toBeInTheDocument()
-    })
-  })
-
-  it('should show the list box on 3 or more character input', async () => {
-    const user = userEvent.setup()
-
-    render(<SideBar />)
-
-    const input = screen.getByRole('combobox', { name: 'Zoek op adres' })
-
-    await user.type(input, 'abc')
-
-    await waitFor(() => {
-      const listBox = screen.getByRole('listbox')
-      expect(listBox).toBeInTheDocument()
-    })
-  })
-
-  it('should show all options returned by the API', async () => {
-    const user = userEvent.setup()
-
-    render(<SideBar />)
-
-    const input = screen.getByRole('combobox', { name: 'Zoek op adres' })
-
-    await user.type(input, 'abc')
-
-    await waitFor(() => {
-      const listItems = screen.getAllByRole('option')
-      expect(listItems).toHaveLength(5)
-
-      expect(listItems[0]).toHaveTextContent('Amsteldijk 152A-H, 1079LG Amsterdam')
-      expect(listItems[1]).toHaveTextContent('Amstelkade 166A-H, 1078AX Amsterdam')
-      expect(listItems[2]).toHaveTextContent('Amstelkade 169A-H, 1078AZ Amsterdam')
-      expect(listItems[3]).toHaveTextContent('Amstelveenseweg 170B-H, 1075XP Amsterdam')
-      expect(listItems[4]).toHaveTextContent('Amstel 312A-H, 1017AP Amsterdam')
-    })
-  })
-
-  it('should show a "no results" message when no results are returned', async () => {
-    server.use(
-      http.get(ENDPOINTS.PDOK_SUGGEST, () =>
-        HttpResponse.json({
-          response: { numFound: 0, start: 0, maxScore: 0.0, numFoundExact: true, docs: [] },
-        }),
-      ),
-    )
-
-    const user = userEvent.setup()
-
-    render(<SideBar />)
-
-    const input = screen.getByRole('combobox', { name: 'Zoek op adres' })
-
-    await user.type(input, 'abc')
-
-    await waitFor(() => {
-      const noResults = screen.getByRole('option', { name: 'Geen resultaten gevonden' })
-      expect(noResults).toBeInTheDocument()
     })
   })
 })
